Check if using XWayland (#8)
The old implementation uses the `XDG_SESSION_TYPE` variable to determine if Wayland is being used. This does not account for users opting to force the usage of X through XWayland, which would be done by unsetting the environment variable `WAYLAND_DISPLAY`. This commit accounts for the unsetting of the `WAYLAND_DISPLAY` variable by falling back to the X11 method.